Move cregbundle carg check to circuit_drawer
